backend/requirements.txt
fastapi>=0.110.0
uvicorn[standard]>=0.27.0
google-adk>=0.1.0
google-cloud-aiplatform>=1.40.0
python-dotenv>=1.0.0
httpx>=0.27.0
pydantic>=2.0.0
polyline>=2.0.0
